added jotai state management
