+++ src/store/chat-state.ts
@@ -1,0 +1,4 @@
+import { INTRO_TEXT } from '@constants'
+import { atom } from 'jotai'
+
+export const aiResponseAtom = atom(INTRO_TEXT)
