+++ tsconfig.json
@@ -5,9 +5,15 @@
     "lib": ["ES2020", "DOM", "DOM.Iterable"],
     "module": "ESNext",
     "skipLibCheck": true,
+    "baseUrl": "./src", // This must be specified if paths are specified
+    "paths": {
+      "*": ["*"],
+      "@store/*": ["store/*"],
+      "@constants": ["constants"]
+    },
 
     /* Bundler mode */
-    "moduleResolution": "bundler",
+    "moduleResolution": "Bundler",
     "allowImportingTsExtensions": true,
     "resolveJsonModule": true,
     "isolatedModules": true,
